ogc tool-support, linting and split up god-file
