+++ backend/pyproject.toml
@@ -53,7 +53,7 @@ isort = "*"
 alembic = "^1.11"
 
 [tool.flake8]
-max-line-length = 100
+max-line-length = 120
 # E203: whitespace before ':' - conflicts with Black formatter for slice notation
 # W503: line break before binary operator - conflicts with Black formatter
 extend-ignore = ["E203", "W503"]
